+++ benchmarks/plugin-punctuation.ts
@@ -1,6 +1,7 @@
 import { bench, run, group, barplot } from 'mitata'
 import MarkdownExit from 'markdown-exit'
-import { markdownItComark } from 'comark/plugins/syntax'
+import { markdownItComponents } from 'comark/plugins/components'
+import { markdownItAttributes } from 'comark/plugins/attributes'
 import { createMarkdownParser } from 'comark'
 import { writeAnsi } from '@comark/ansi'
 import punctuation from '../packages/comark/src/plugins/punctuation'
@@ -46,7 +47,8 @@ const parserTypographer = new MarkdownExit({
   typographer: true,
 })
   .enable(['table', 'strikethrough'])
-  .use(markdownItComark)
+  .use(markdownItComponents)
+  .use(markdownItAttributes)
 
 const parserNoTypographer = new MarkdownExit({
   html: false,
@@ -54,7 +56,8 @@ const parserNoTypographer = new MarkdownExit({
   typographer: false,
 })
   .enable(['table', 'strikethrough'])
-  .use(markdownItComark)
+  .use(markdownItComponents)
+  .use(markdownItAttributes)
 
 // ── comark with full punctuation plugin (all features) ──────────────────────
