+++ docs/content/2.syntax/2.components.md
@@ -7,20 +7,19 @@ links:
     to: /syntax/attributes
     color: neutral
     variant: soft
-  - label: Document Model
-    icon: i-lucide-braces
-    to: /getting-started/document-model
+  - label: Components Plugin
+    icon: i-lucide-plug
+    to: /plugins/built-in/components
     color: neutral
     variant: soft
 navigation:
   title: Components
 ---
 
-Comark extends standard Markdown with component syntax, letting you embed custom UI directly in your content while keeping it readable.
-
+Comark extends standard Markdown with component syntax, letting you embed custom UI directly in your content while keeping it readable. Parsing is provided by the built-in [`components` plugin](/plugins/built-in/components) (enabled by default).
 
 ::callout
-Comark also support JSON-render syntax in code-blocks to render components, checkout the [json-render plugin](http://localhost:3000/plugins/built-in/json-render).
+Comark also supports JSON-render syntax in code blocks — see the [json-render plugin](/plugins/built-in/json-render).
 ::
 
 ## Block
