+++ docs/content/2.syntax/3.attributes.md
@@ -2,21 +2,21 @@
 title: Attributes
 description: Add custom classes, IDs, styles, and data attributes to native Markdown elements using the curly brace syntax.
 links:
-  - label: Markdown
-    icon: i-lucide-file-text
-    to: /syntax/markdown
-    color: neutral
-    variant: soft
   - label: Components
     icon: i-lucide-component
     to: /syntax/components
+    color: neutral
+    variant: soft
+  - label: Attributes Plugin
+    icon: i-lucide-plug
+    to: /plugins/built-in/attributes
     color: neutral
     variant: soft
 navigation:
   title: Attributes
 ---
 
-Comark allows adding custom attributes to native Markdown elements using `{...}` syntax after the element.
+Comark allows adding custom attributes to native Markdown elements using `{...}` syntax after the element. Parsing is provided by the built-in [`attributes` plugin](/plugins/built-in/attributes) (enabled by default).
 
 ## Element Attributes
