+++ docs/content/5.api/1.parse.md
@@ -166,7 +166,7 @@ console.log(result.nodes)
 ```
 ::
 
-To disable HTML parsing and treat tags as plain text, set `html: false`:
+HTML parsing is provided by the built-in [`html` plugin](/plugins/built-in/html). To disable it and treat tags as plain text, set `html: false`:
 
 ```typescript [parse.ts]
 const result = await parseMarkdown(content, { html: false })
@@ -364,11 +364,11 @@ Both `parseMarkdown()` and `createMarkdownParser()` accept the same `ParserOptio
 | `autoUnwrap` | `boolean` | `true` | Remove unnecessary `<p>` wrappers from single-element containers |
 | `autoClose` | `boolean` | `true` | Auto-close incomplete markdown syntax |
 | `unwrap` | `boolean \| string \| string[]` | `false` | Remove wrapper tags from the tree, hoisting their children (MDC `unwrap` behaviour). `true` unwraps `p`; a comma/whitespace-separated string or array unwraps the listed tags; `'*'` matches any tag. Tags apply sequentially (each descends one level), and adjacent text is merged into a single string. |
-| `html` | `boolean` | `true` | Parse embedded HTML tags into AST nodes |
+| `html` | `boolean` | `true` | Include the built-in [`html` plugin](/plugins/built-in/html) among defaults (parse embedded HTML tags into AST nodes). Only applies when `registerDefaultPlugins` is not `false` |
 | `linkify` | `boolean` | `true` | Auto-convert URL-like text into links. Set `false` to disable |
 | `headingIds` | `boolean` | `true` | Auto-generate `id` attributes for `h1`–`h6` headings. Set `false` to disable |
 | `frontmatter` | `boolean` | `true` | Parse a leading YAML frontmatter block into `tree.frontmatter`. Set `false` to treat the `---` block as regular markdown content |
-| `registerDefaultPlugins` | `boolean` | `true` | Register the built-in default plugins (`alert`, `task-list`, `syntax`). Set `false` to disable. Also can be used to configure default plugins like `syntax` in conjunction with `plugins` |
+| `registerDefaultPlugins` | `boolean` | `true` | Register the built-in default plugins (`html`, `alert`, `task-list`, `components`, `attributes`). Set `false` to disable. Also can be used to configure default plugins like `components` in conjunction with `plugins` |
 | `plugins` | `ComarkPlugin[]` | `[]` | Array of plugins to apply |
 
 ### Inline rendering
