+++ docs/content/5.api/2.auto-close.md
@@ -21,7 +21,7 @@ Automatically closes unclosed markdown inline syntax and Comark components. Buil
 **Parameters:**
 
 - `source` - The markdown content (potentially partial/incomplete)
-- `options` (optional) - `frontmatter: true` completes an unclosed leading frontmatter block; `syntax: false` skips Comark component-fence closing (for content parsed without the syntax plugin)
+- `options` (optional) - `frontmatter: true` completes an unclosed leading frontmatter block; `syntax: false` skips Comark component-fence closing (for content parsed without the components plugin)
 
 **Returns:** `string`, the source with all unclosed syntax closed
