+++ packages/comark/src/plugins/components.ts
@@ -1,57 +1,19 @@
-import type { MarkdownExit, PluginSimple, Renderer } from 'markdown-exit'
-import { Token } from 'markdown-exit'
-import type { MarkdownItPlugin, MarkdownItPluginWithOptions } from '../types.ts'
+/**
+ * Components syntax plugin for Comark.
+ *
+ * Enables block (`::name`) and inline (`:name[content]`) component syntax,
+ * plus span wrappers (`[text]` / `[text]{attrs}` when paired with attributes).
+ * On by default via `registerDefaultPlugins`.
+ *
+ * @see https://comark.dev/syntax/components
+ */
+
+import type { PluginSimple, Token } from 'markdown-exit'
+import type { MarkdownItPlugin } from '../types.ts'
 import { defineComarkPlugin } from '../utils/helpers.ts'
 import { findClosingBracket, parseBracketContent } from '../internal/parse/syntax/brackets.ts'
-import { searchProps } from '../internal/parse/syntax/props.ts'
 import { parseBlockParams } from '../internal/parse/syntax/block-params.ts'
 import { parseYaml } from '../internal/yaml.ts'
-
-export interface SyntaxOptions {
-  /**
-   * Enable block component syntax.
-   *
-   * @see https://comark.dev/syntax/components#block
-   * @default true
-   */
-  blockComponent?: boolean
-  /**
-   * Enable inline props syntax.
-   *
-   * @see https://comark.dev/syntax/attributes
-   * @default true
-   */
-  inlineProps?: boolean
-  /**
-   * Enable inline span syntax.
-   *
-   * @see https://comark.dev/syntax/attributes#span-attributes
-   * @default true
-   */
-  inlineSpan?: boolean
-  /**
-   * Enable inline component syntax.
-   *
-   * @see https://comark.dev/syntax/components#inline
-   * @default true
-   */
-  inlineComponent?: boolean
-  /**
-   * Enable inline binding syntax (`{{ value }}` and `{{ value || default }}`).
-   *
-   * Off by default — opt in here, or use the standalone `plugins/binding.ts` plugin.
-   *
-   * @see https://comark.dev/syntax/components#data-binding
-   * @default false
-   */
-  inlineBinding?: boolean
-  /**
-   * The tag name used to render an inline binding.
-   *
-   * @default 'binding'
-   */
-  bindingTag?: string
-}
 
 /**
  * A component name must start with a letter or `$`, followed by word chars,
@@ -70,8 +32,6 @@ const RE_COMPONENT_NAME = /^[a-z$][\w$-]*/i
 function isValidComponentName(name: string): boolean {
   return RE_COMPONENT_NAME.test(name)
 }
-
-// #region Block component plugin (`::name` and `::name ... ::`)
 
 const blockYamlLines: Record<string, string> = {
   '---': '---',
@@ -416,52 +376,6 @@ const markdownItComarkBlock: PluginSimple = (md) => {
   })
 }
 
-// #endregion
-
-// #region Inline span plugin (`[text]`)
-
-const markdownItInlineSpan: PluginSimple = (md) => {
-  md.inline.ruler.before('link', 'comark_inline_span', (state, silent) => {
-    const start = state.pos
-    if (state.src[start] !== '[') return false
-
-    // An unclosed span consumes to the end of input (streaming auto-close)
-    const close = findClosingBracket(state.src, start)
-    const index = close === -1 ? state.src.length : close
-
-    // Don't match `[text](url)` or `[text][ref]` — let the link parser handle those
-    const nextChar = state.src[index + 1]
-    if (nextChar === '(' || nextChar === '[') return false
-
-    // Inside a link label, bare `[text]` stays literal (plain-markdown behavior,
-    // e.g. `[[1] Document](#)`); only an explicit `[text]{attrs}` span matches
-    if (state.linkLevel > 0 && nextChar !== '{') return false
-
-    // Returning `false` lets `parseLinkLabel`'s own depth tracking consume nested brackets and the outer link parse
-    if (silent) return false
-
-    state.push('mdc_inline_span', 'span', 1)
-
-    const oldPos = state.pos
-    const oldPosMax = state.posMax
-    state.pos = start + 1
-    state.posMax = index
-    state.md.inline.tokenize(state)
-    state.pos = oldPos
-    state.posMax = oldPosMax
-
-    state.push('mdc_inline_span', 'span', -1)
-
-    state.pos = index + 1
-
-    return true
-  })
-}
-
-// #endregion
-
-// #region Inline component plugin (`:name[content]{props}`)
-
 const ALLOWED_PREV_CHARS = new Set([' ', '\t', '\n', '*', '_', '['])
 
 const markdownItInlineComponent: PluginSimple = (md) => {
@@ -525,168 +439,56 @@ const markdownItInlineComponent: PluginSimple = (md) => {
   })
 }
 
-// #endregion
+// #region Inline span (`[text]` / `[text]{attrs}`)
 
-// #region Inline props plugin (`{class="foo"}` after a token)
-
-const markdownItInlineProps: PluginSimple = (md) => {
-  md.inline.ruler.after('entity', 'comark_inline_props', (state, silent) => {
+const markdownItInlineSpan: PluginSimple = (md) => {
+  md.inline.ruler.before('link', 'comark_inline_span', (state, silent) => {
     const start = state.pos
-    if (state.src[start] !== '{') return false
+    if (state.src[start] !== '[') return false
 
-    // Skip Vue mustache `{{ }}` and template `${ }` syntax
-    if (state.src[start + 1] === '{' || state.src[start - 1] === '{' || state.src[start - 1] === '$') return false
+    // An unclosed span consumes to the end of input (streaming auto-close)
+    const close = findClosingBracket(state.src, start)
+    const index = close === -1 ? state.src.length : close
 
-    const search = searchProps(state.src, start)
-    if (!search) return false
+    // Don't match `[text](url)` or `[text][ref]` — let the link parser handle those
+    const nextChar = state.src[index + 1]
+    if (nextChar === '(' || nextChar === '[') return false
 
-    const { props, index: end } = search
-    if (end === start) return false
+    // Inside a link label, bare `[text]` stays literal (plain-markdown behavior,
+    // e.g. `[[1] Document](#)`); only an explicit `[text]{attrs}` span matches
+    if (state.linkLevel > 0 && nextChar !== '{') return false
 
-    state.pos = end
+    // Returning `false` lets `parseLinkLabel`'s own depth tracking consume nested brackets and the outer link parse
+    if (silent) return false
 
-    if (silent) return true
+    state.push('mdc_inline_span', 'span', 1)
 
-    // Hidden token holding the props; later applied to the previous token
-    const token = state.push('mdc_inline_props', 'span', 0)
-    token.attrs = props
-    token.hidden = true
+    const oldPos = state.pos
+    const oldPosMax = state.posMax
+    state.pos = start + 1
+    state.posMax = index
+    state.md.inline.tokenize(state)
+    state.pos = oldPos
+    state.posMax = oldPosMax
+
+    state.push('mdc_inline_span', 'span', -1)
+
+    state.pos = index + 1
 
     return true
   })
-
-  md.renderer.rules.mdc_inline_props = () => ''
-
-  const _parse = md.parse
-  md.parse = function (src, env) {
-    const tokens = _parse.call(this, src, env)
-
-    // When the trailing inline child is a props token directly after a text
-    // node, lift the props onto the surrounding heading/paragraph/list_item.
-    // (If the props follow a closing tag, they belong to that inline tag, not
-    // the parent — leave them alone.)
-    tokens.forEach((token, index) => {
-      const prev = tokens[index - 1]
-      const next = tokens[index + 1]
-      if (!prev || !['heading_open', 'paragraph_open', 'list_item_open'].includes(prev.type) || prev.hidden) return
-
-      // Tight-list paragraph: the inline lives one slot ahead
-      if (token.hidden && next?.type === 'inline') token = next
-
-      if (token.type !== 'inline' || !token.children?.length) return
-
-      const last = token.children[token.children.length - 1]
-      if (last.type !== 'mdc_inline_props') return
-
-      // Find the previous non-empty child. Markdown-it's emphasis tokenizer
-      // can leave an empty text token between the closing delimiter and the
-      // props — skipping it lets us distinguish "props on the parent" from
-      // "props on the preceding inline tag".
-      let beforeIdx = token.children.length - 2
-      while (beforeIdx >= 0) {
-        const child = token.children[beforeIdx]
-        if (child.type === 'text' && !child.content) {
-          beforeIdx--
-          continue
-        }
-        break
-      }
-      const beforeProps = beforeIdx >= 0 ? token.children[beforeIdx] : undefined
-      if (!beforeProps || beforeProps.type !== 'text') return
-
-      // Strip the trailing space the text picked up before the `{...}` token.
-      if (typeof beforeProps.content === 'string') {
-        beforeProps.content = beforeProps.content.replace(/[ \t]+$/, '')
-      }
-
-      const props = last.attrs
-      // Drop the props token (last) plus any empty text tokens it left behind.
-      token.children.length = beforeProps.content ? beforeIdx + 1 : beforeIdx
-      props?.forEach(([key, value]) => {
-        if (key === 'class') prev.attrJoin('class', value)
-        else prev.attrSet(key, value)
-      })
-    })
-
-    return tokens
-  }
-
-  md.renderer.renderInline = wrapRenderInline(md.renderer.renderInline)
-  // Support markdown-exit's async inline renderer
-  if ('renderInlineAsync' in md.renderer) {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any -- mirrors the sync wrapper for the async overload
-    ;(md.renderer as any).renderInlineAsync = wrapRenderInline((md.renderer as any).renderInlineAsync)
-  }
-}
-
-function wrapRenderInline(renderInline: Renderer['renderInline']): Renderer['renderInline'] {
-  return function (this: Renderer, tokens, options, env) {
-    tokens = [...tokens]
-    tokens.forEach((token, index) => {
-      if (token.type !== 'mdc_inline_props') return
-
-      let prevIndex = index - 1
-      let prev = tokens[prevIndex]
-      // Skip whitespace-only text tokens
-      while (prevIndex >= 0) {
-        if (prev.type === 'text' && !prev.content.trim()) {
-          prevIndex--
-          prev = tokens[prevIndex]
-        } else {
-          break
-        }
-      }
-
-      // Wrap a bare text token in a span so we can attach attrs to it
-      if (!prev.tag && prev.type === 'text') {
-        prev = new Token('mdc_inline_span', 'span', 1)
-        tokens.splice(index - 1, 0, prev)
-        const close = new Token('mdc_inline_span', 'span', -1)
-        tokens.splice(index + 2, 0, close)
-      } else if (prev.nesting === -1) {
-        // Resolve a closing tag back to its matching opening tag
-        let searchIndex = index - 1
-        while (searchIndex >= 0) {
-          const searchToken = tokens[searchIndex]
-          if (searchToken.nesting === 1 && searchToken.tag === prev.tag && searchToken.level === prev.level) {
-            prev = searchToken
-            break
-          }
-          searchIndex--
-        }
-      }
-
-      if (prev.nesting === -1) throw new Error(`No matching opening tag found for ${JSON.stringify(prev)}`)
-
-      token.attrs?.forEach(([key, value]) => {
-        if (key === 'class') prev.attrJoin('class', value)
-        else prev.attrSet(key, value)
-      })
-    })
-    return renderInline.call(this, tokens, options, env)
-  }
 }
 
 // #endregion
 
-function applySyntax(md: MarkdownExit, options: SyntaxOptions = {}) {
-  const { blockComponent = true, inlineProps = true, inlineSpan = true, inlineComponent = true } = options
-
-  if (blockComponent) md.use(markdownItComarkBlock)
-  if (inlineProps) md.use(markdownItInlineProps)
-  if (inlineSpan) md.use(markdownItInlineSpan)
-  if (inlineComponent) md.use(markdownItInlineComponent)
+/** markdown-it / markdown-exit adapter for Comark component + span syntax. */
+export const markdownItComponents: PluginSimple = (md) => {
+  md.use(markdownItComarkBlock)
+  md.use(markdownItInlineSpan)
+  md.use(markdownItInlineComponent)
 }
 
-export class SyntaxPlugin {
-  name = 'syntax'
-  markdownItPlugins: MarkdownItPlugin[]
-
-  constructor(options: SyntaxOptions = {}) {
-    this.markdownItPlugins = [((md: MarkdownExit) => applySyntax(md, options)) as unknown as MarkdownItPlugin]
-  }
-}
-
-export default defineComarkPlugin((options: SyntaxOptions = {}) => new SyntaxPlugin(options))
-
-export const markdownItComark = applySyntax as unknown as MarkdownItPluginWithOptions<SyntaxOptions>
+export default defineComarkPlugin(() => ({
+  name: 'components',
+  markdownItPlugins: [markdownItComponents as unknown as MarkdownItPlugin],
+}))
