+++ packages/comark/test/plugins/default-plugins.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest'
 import { parseMarkdown } from '../../src/parse'
-import syntax from '../../src/plugins/syntax'
+import components from '../../src/plugins/components'
+import html from '../../src/plugins/html'
 
 describe('default plugin options', () => {
   describe('registerDefaultPlugins', () => {
@@ -17,6 +18,34 @@ describe('default plugin options', () => {
     it('transforms alert blockquotes by default', async () => {
       const tree = await parseMarkdown('> [!NOTE]\n> hi')
       expect(tree.nodes).toEqual([['blockquote', { as: 'note' }, 'hi']])
+    })
+
+    it('parses HTML by default', async () => {
+      const tree = await parseMarkdown('<strong class="bold">Hello</strong>')
+      expect(tree.nodes).toEqual([['p', {}, ['strong', { class: 'bold', $: { html: 1, block: 0 } }, 'Hello']]])
+    })
+
+    it('treats HTML as plain text when registerDefaultPlugins is false', async () => {
+      const tree = await parseMarkdown('<em>hi</em>', { registerDefaultPlugins: false })
+      expect(tree.nodes).toEqual([['p', {}, '<em>hi</em>']])
+    })
+
+    it('parses HTML via an explicit html plugin when registerDefaultPlugins is false', async () => {
+      const tree = await parseMarkdown('<em>hi</em>', {
+        registerDefaultPlugins: false,
+        plugins: [html()],
+      })
+      expect(tree.nodes).toEqual([['p', {}, ['em', { $: { html: 1, block: 0 } }, 'hi']]])
+    })
+
+    it('parses attributes by default', async () => {
+      const tree = await parseMarkdown('Hello {.cls}')
+      expect(tree.nodes).toEqual([['p', { class: 'cls' }, 'Hello']])
+    })
+
+    it('treats attribute braces as plain text when registerDefaultPlugins is false', async () => {
+      const tree = await parseMarkdown('Hello {.cls}', { registerDefaultPlugins: false })
+      expect(tree.nodes).toEqual([['p', {}, 'Hello {.cls}']])
     })
 
     it('leaves component syntax literal when disabled', async () => {
@@ -53,12 +82,41 @@ describe('default plugin options', () => {
     })
   })
 
-  describe('user plugin override', () => {
-    it('configures the syntax plugin through plugins', async () => {
-      const tree = await parseMarkdown('Hello [world]{.cls}', {
-        plugins: [syntax({ inlineSpan: false })],
+  describe('html option', () => {
+    it('treats HTML as plain text when html: false', async () => {
+      const tree = await parseMarkdown('<strong>Hello</strong>', { html: false })
+      expect(tree.nodes).toEqual([['p', {}, '<strong>Hello</strong>']])
+    })
+
+    it('enables HTML via the explicit html plugin when html: false', async () => {
+      const tree = await parseMarkdown('<strong>Hello</strong>', {
+        html: false,
+        plugins: [html()],
       })
-      expect(tree.nodes).toEqual([['p', { class: 'cls' }, 'Hello [world]']])
+      expect(tree.nodes).toEqual([['p', {}, ['strong', { $: { html: 1, block: 0 } }, 'Hello']]])
+    })
+
+    it('replaces the default html plugin with a same-name user plugin', async () => {
+      const tree = await parseMarkdown('<strong>Hello</strong>', {
+        plugins: [{ name: 'html' }],
+      })
+      expect(tree.nodes).toEqual([['p', {}, '<strong>Hello</strong>']])
+    })
+
+    it('disables HTML via html({ enabled: false })', async () => {
+      const tree = await parseMarkdown('<strong>Hello</strong>', {
+        plugins: [html({ enabled: false })],
+      })
+      expect(tree.nodes).toEqual([['p', {}, '<strong>Hello</strong>']])
+    })
+  })
+
+  describe('user plugin override', () => {
+    it('replaces the default attributes plugin with a same-name user plugin', async () => {
+      const tree = await parseMarkdown('Hello {.cls}', {
+        plugins: [{ name: 'attributes' }],
+      })
+      expect(tree.nodes).toEqual([['p', {}, 'Hello {.cls}']])
     })
 
     it('replaces the default alert plugin with a same-name user plugin', async () => {
@@ -75,17 +133,17 @@ describe('default plugin options', () => {
       expect(tree.nodes).toEqual([['ul', {}, ['li', {}, ' todo']]])
     })
 
-    it('keeps an explicit syntax plugin active with registerDefaultPlugins: false', async () => {
+    it('keeps an explicit components plugin active with registerDefaultPlugins: false', async () => {
       const tree = await parseMarkdown('::alert\nContent', {
         registerDefaultPlugins: false,
-        plugins: [syntax()],
+        plugins: [components()],
       })
       expect(tree.nodes).toEqual([['alert', {}, 'Content']])
     })
 
-    it('skips component-fence auto-close when a custom plugin replaces syntax', async () => {
+    it('skips component-fence auto-close when components is replaced', async () => {
       const tree = await parseMarkdown('::alert\nContent', {
-        plugins: [{ name: 'syntax' }],
+        plugins: [{ name: 'components' }],
       })
       expect(tree.nodes).toEqual([['p', {}, '::alert\nContent']])
     })
